pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "inforoutine"
version = "1.0.0"
description = "Strategic Intelligence Daily Briefing — AI-powered news synthesis for Notion"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.0.0",
    "notion-client>=2.2.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "openpyxl>=3.1.0",
    "feedparser>=6.0.11",
    "beautifulsoup4>=4.12.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py312"
line-length = 100
